refactor: extract PDF conversion rendering
